Phase I: extract pure-compute offline rewrite into offline-rewrite.mjs
